Change names of some directories in test/
- cli_command_test/ -> cli_command/
- rhsmlib_test/ -> rhsmlib/

There is no need to keep the '_test' suffix, these directories are
inside of directory named 'test/'.
