+++ test/cli_command/test_addons.py
