+++ test/cli_command/test_attach.py
