+++ test/cli_command/test_clean.py
