+++ test/cli_command/test_environments.py
