+++ test/cli_command/test_list.py
