+++ test/cli_command/test_plugins.py
