+++ test/cli_command/test_redeem.py
