+++ test/cli_command/test_register.py
