+++ test/cli_command/test_remove.py
