+++ test/cli_command/test_role.py
