+++ test/cli_command/test_status.py
