+++ test/cli_command/test_unregister.py
