+++ test/rhsmlib/dbus/test_attach.py
@@ -14,7 +14,7 @@ import dbus
 import json
 import mock
 
-from test.rhsmlib_test.base import DBusObjectTest, InjectionMockingTest
+from test.rhsmlib.base import DBusObjectTest, InjectionMockingTest
 
 from subscription_manager import injection as inj
 from subscription_manager.cp_provider import CPProvider
