+++ test/rhsmlib/facts/test_cloud_facts.py
