+++ test/rhsmlib/facts/test_collector.py
