+++ test/rhsmlib/facts/test_host_collector.py
