+++ test/rhsmlib/facts/test_insights.py
