+++ test/rhsmlib/facts/test_kpatch.py
