+++ test/rhsmlib/facts/test_pkg_arches.py
