+++ test/rhsmlib/services/test_attach.py
@@ -12,7 +12,7 @@
 # in this software or its documentation.
 import mock
 
-from test.rhsmlib_test.base import InjectionMockingTest
+from test.rhsmlib.base import InjectionMockingTest
 
 from subscription_manager import injection as inj
 from subscription_manager.identity import Identity
