+++ test/rhsmlib/services/test_consumer.py
@@ -17,7 +17,7 @@ from rhsmlib.services import consumer
 
 from subscription_manager import injection as inj
 from subscription_manager.identity import Identity
-from test.rhsmlib_test.base import InjectionMockingTest
+from test.rhsmlib.base import InjectionMockingTest
 
 
 class TestConsumerService(InjectionMockingTest):
