+++ test/rhsmlib/services/test_unregister.py
@@ -14,7 +14,7 @@
 
 import mock
 
-from test.rhsmlib_test.base import InjectionMockingTest
+from test.rhsmlib.base import InjectionMockingTest
 
 from subscription_manager import injection as inj
 from subscription_manager.identity import Identity
