+++ test/zypper/test_serviceplugin.py
